benerin bug sonarqube

+++ rumahsehat/rumahsehat/src/main/java/apap/proyek/rumahsehat/security/jwt_config/JwtTokenUtil.java
@@ -23,7 +23,7 @@ public class JwtTokenUtil implements Serializable {
 
 	private static final long serialVersionUID = -2550185165626007488L;
 
-	public static final long JWT_TOKEN_VALIDITY = 60*60*60;
+	public static final long JWT_TOKEN_VALIDITY = (long) 60*60*60;
 
 	@Value("proyekapaprumahsehatapap50springbootflutter")
 	private String secret;
